+++ src/main/java/com/bugbycode/controller/user/UserController.java
@@ -10,10 +10,16 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.alibaba.fastjson.JSONObject;
+import com.bugbycode.module.role.Role;
 import com.bugbycode.module.user.User;
+import com.bugbycode.module.user.UserGroup;
+import com.bugbycode.service.role.RoleService;
+import com.bugbycode.service.user.UserGroupService;
 import com.bugbycode.service.user.UserService;
+import com.util.AESUtil;
 import com.util.StringUtil;
 import com.util.page.SearchResult;
+import com.util.reg.RegexUtil;
 
 /**
  * 用户信息管理API
@@ -27,6 +33,12 @@ public class UserController {
 	
 	@Autowired
 	private UserService userService;
+	
+	@Autowired
+	private RoleService roleService;
+	
+	@Autowired
+	private UserGroupService userGroupService;
 	
 	/**
 	 * 自定义条件查询用户信息
@@ -96,10 +108,37 @@ public class UserController {
 	@ResponseBody
 	public String update(String jsonStr) {
 		User user = JSONObject.toJavaObject(JSONObject.parseObject(jsonStr), User.class);
-		userService.update(user);
+		int userId = user.getId();
+		String name = user.getName();
+		String email = user.getEmail();
+		String password = user.getPassword();
+		int code = 0;
+		String msg = "修改成功";
+		try {
+			if(userId <= 0) {
+				throw new RuntimeException("用户ID信息错误");
+			}
+			if(!RegexUtil.check(RegexUtil.USER_NAME_REGEX, name)) {
+				throw new RuntimeException("用户姓名格式错误");
+			}
+			if(StringUtil.isNotBlank(email) && !RegexUtil.check(RegexUtil.EMAIL_REGEX, email)) {
+				throw new RuntimeException("电子邮件格式错误");
+			}
+			User old = userService.queryByUserId(userId);
+			if(old == null) {
+				throw new RuntimeException("该用户不存在");
+			}
+			if(StringUtil.isNotEmpty(password)) {
+				user.setPassword(AESUtil.encrypt(password));
+			}
+			userService.update(user);
+		}catch (Exception e) {
+			code = 1;
+			msg = e.getMessage();
+		}
 		JSONObject json = new JSONObject();
-		json.put("msg", "修改成功");
-		json.put("code", 0);
+		json.put("msg", msg);
+		json.put("code", code);
 		return json.toJSONString();
 	}
 	
@@ -112,11 +151,40 @@ public class UserController {
 	@ResponseBody
 	public String insert(String jsonStr) {
 		User user = JSONObject.toJavaObject(JSONObject.parseObject(jsonStr), User.class);
-		userService.insert(user);
+		String name = user.getName();
+		String username = user.getUsername();
+		String password = user.getPassword();
+		String email = user.getEmail();
+		int code = 0;
+		String msg = "新建成功";
+		try {
+			if(!RegexUtil.check(RegexUtil.USER_NAME_REGEX, name)) {
+				throw new RuntimeException("用户姓名格式错误");
+			}
+			if(StringUtil.isNotBlank(password)) {
+				user.setPassword(AESUtil.encrypt(password));
+			}
+			if(!RegexUtil.check(RegexUtil.USER_LOGIN_NAME_REGEX, username)) {
+				throw new RuntimeException("用户名格式错误");
+			}
+			if(StringUtil.isNotBlank(email) && !RegexUtil.check(RegexUtil.EMAIL_REGEX, email)) {
+				throw new RuntimeException("电子邮件格式错误");
+			}
+			User old = userService.queryByUserName(username);
+			if(old != null) {
+				throw new RuntimeException("该用户名已被使用");
+			}
+			user.setType(1);
+			userService.insert(user);
+		}catch (Exception e) {
+			code = 1;
+			msg = e.getMessage();
+		}
+		int userId = user.getId();
 		JSONObject json = new JSONObject();
-		json.put("msg", "新建成功");
-		json.put("userId", user.getId());
-		json.put("code", 0);
+		json.put("msg", msg);
+		json.put("userId", userId);
+		json.put("code", code);
 		return json.toJSONString();
 	}
 	
@@ -128,10 +196,24 @@ public class UserController {
 	@RequestMapping("/delete")
 	@ResponseBody
 	public String delete(int userId) {
-		userService.delete(userId);
+		int code = 0;
+		String msg = "删除成功";
+		try {
+			if(userId <= 0) {
+				throw new RuntimeException("用户ID信息错误");
+			}
+			User user = userService.queryByUserId(userId);
+			if(user == null) {
+				throw new RuntimeException("该用户信息不存在");
+			}
+			userService.delete(userId);
+		}catch (Exception e) {
+			msg = e.getMessage();
+			code = 1;
+		}
 		JSONObject json = new JSONObject();
-		json.put("msg", "删除成功");
-		json.put("code", 0);
+		json.put("msg", msg);
+		json.put("code", code);
 		return json.toJSONString();
 	}
 	
@@ -144,10 +226,31 @@ public class UserController {
 	@RequestMapping("/insertRelRole")
 	@ResponseBody
 	public String insertRelRole(int userId,int roleId) {
-		userService.insertRelRole(userId, roleId);
+		int code = 0;
+		String msg = "新建成功";
+		try {
+			if(userId <= 0) {
+				throw new RuntimeException("用户ID错误");
+			}
+			User user = userService.queryByUserId(userId);
+			if(user == null) {
+				throw new RuntimeException("用户ID错误");
+			}
+			if(roleId <= 0) {
+				throw new RuntimeException("角色ID错误");
+			}
+			Role role = roleService.queryByRoleId(roleId);
+			if(role == null) {
+				throw new RuntimeException("角色ID错误");
+			}
+			userService.insertRelRole(userId, roleId);
+		}catch (Exception e) {
+			code = 1;
+			msg = e.getMessage();
+		}
 		JSONObject json = new JSONObject();
-		json.put("msg", "新建成功");
-		json.put("code", 0);
+		json.put("msg", msg);
+		json.put("code", code);
 		return json.toJSONString();
 	}
 	
@@ -159,10 +262,24 @@ public class UserController {
 	@RequestMapping("/deleteRelRoleByUserId")
 	@ResponseBody
 	public String deleteRelRoleByUserId(int userId) {
-		userService.deleteRelRoleByUserId(userId);
+		int code = 0;
+		String msg = "删除成功";
+		try {
+			if(userId < 1) {
+				throw new RuntimeException("用户ID错误");
+			}
+			User user = userService.queryByUserId(userId);
+			if(user == null) {
+				throw new RuntimeException("用户ID错误");
+			}
+			userService.deleteRelRoleByUserId(userId);
+		}catch (Exception e) {
+			code = 1;
+			msg = e.getMessage();
+		}
 		JSONObject json = new JSONObject();
-		json.put("msg", "删除成功");
-		json.put("code", 0);
+		json.put("msg", msg);
+		json.put("code", code);
 		return json.toJSONString();
 	}
 	
@@ -175,10 +292,31 @@ public class UserController {
 	@RequestMapping("/insertRelGroup")
 	@ResponseBody
 	public String insertRelGroup(int userId,int groupId) {
-		userService.insertRelGroup(userId, groupId);
+		int code = 0;
+		String msg = "删除成功";
+		try {
+			if(userId < 1) {
+				throw new RuntimeException("用户ID错误");
+			}
+			User user = userService.queryByUserId(userId);
+			if(user == null) {
+				throw new RuntimeException("用户ID错误");
+			}
+			if(groupId < 1) {
+				throw new RuntimeException("用户分组ID错误");
+			}
+			UserGroup group = userGroupService.queryByGroupId(groupId);
+			if(group == null) {
+				throw new RuntimeException("用户分组ID错误");
+			}
+			userService.insertRelGroup(userId, groupId);
+		}catch (Exception e) {
+			code = 1;
+			msg = e.getMessage();
+		}
 		JSONObject json = new JSONObject();
-		json.put("msg", "新建成功");
-		json.put("code", 0);
+		json.put("msg", msg);
+		json.put("code", code);
 		return json.toJSONString();
 	}
 	
@@ -190,10 +328,24 @@ public class UserController {
 	@RequestMapping("/deleteRelGroupByUserId")
 	@ResponseBody
 	public String deleteRelGroupByUserId(int userId) {
-		userService.deleteRelGroupByUserId(userId);
+		int code = 0;
+		String msg = "删除成功";
+		try {
+			if(userId < 1) {
+				throw new RuntimeException("用户ID错误");
+			}
+			User user = userService.queryByUserId(userId);
+			if(user == null) {
+				throw new RuntimeException("用户ID错误");
+			}
+			userService.deleteRelGroupByUserId(userId);
+		}catch (Exception e) {
+			code = 1;
+			msg = e.getMessage();
+		}
 		JSONObject json = new JSONObject();
-		json.put("msg", "删除成功");
-		json.put("code", 0);
+		json.put("msg", msg);
+		json.put("code", code);
 		return json.toJSONString();
 	}
 }
